+++ src/main/java/ru/yandex/practicum/filmorate/service/db/InDBUserService.java
@@ -6,6 +6,7 @@ import org.springframework.stereotype.Service;
 import ru.yandex.practicum.filmorate.dal.UserRepository;
 import ru.yandex.practicum.filmorate.dto.UserDto;
 import ru.yandex.practicum.filmorate.exception.DuplicatedDataException;
+import ru.yandex.practicum.filmorate.exception.NotFoundException;
 import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.mapper.UserMapper;
 import ru.yandex.practicum.filmorate.model.User;
@@ -33,8 +34,11 @@ public class InDBUserService implements UserService {
 
     @Override
     public UserDto findById(int userId) {
-        UserDB userDb = userRepository.getUserById(userId);
-        User user = mapToUser(userDb);
+        if (userRepository.getUserById(userId).isEmpty())
+            throw new NotFoundException(String.format("Пользователь id = %d не найден", userId));
+        UserDB userDB = userRepository.getUserById(userId).get();
+        userDB.setFriends(userRepository.getAllFriendOfUser(userId));
+        User user = mapToUser(userDB);
         return UserMapper.mapToUserDto(user);
     }
 
@@ -56,10 +60,14 @@ public class InDBUserService implements UserService {
             throw new ValidationException("При обновлении пользователя не указан id");
         }
         int userId = userFromRequest.getId();
+        if (userRepository.getUserById(userId).isEmpty()) {
+            throw new NotFoundException(String.format("Пользователь id = %d не найден", userId));
+        }
         if (checkDuplicatedEmail(userFromRequest)) {
             throw new DuplicatedDataException("Данный имейл уже используется");
         }
-        UserDB oldUser = userRepository.getUserById(userId);
+        UserDB oldUser = userRepository.getUserById(userId).get();
+        oldUser.setFriends(userRepository.getAllFriendOfUser(userId));
         UserDB updatedOldUser = UserMapper.updateUserFields(oldUser, userFromRequest);
         User user = mapToUser(updatedOldUser);
         UserDB userDB = UserMapper.mapToUserDB(user);
@@ -76,7 +84,9 @@ public class InDBUserService implements UserService {
 
     @Override
     public String remove(int userId) {
-        userRepository.getUserById(userId);
+        if (userRepository.getUserById(userId).isEmpty()) {
+            return String.format("Пользователь %d не существует", userId);
+        }
         userRepository.removeUser(userId);
         return String.format("Пользователь id = %d удалён", userId);
     }
@@ -94,25 +104,47 @@ public class InDBUserService implements UserService {
 
     @Override
     public void addFriend(int userId, int friendId) {
-        UserDB user = userRepository.getUserById(userId);
-        userRepository.getUserById(friendId);
+        if (userRepository.getUserById(userId).isEmpty()) {
+            throw new NotFoundException(String.format("Пользователь id = %d не найден", userId));
+        }
+        UserDB user = userRepository.getUserById(userId).get();
+        if (userRepository.getUserById(friendId).isEmpty()) {
+            throw new NotFoundException(String.format("Пользователь id = %d не найден", friendId));
+        }
+        user.setFriends(userRepository.getAllFriendOfUser(userId));
         userRepository.addFriend(userId, friendId);
         user.getFriends().add(friendId);
     }
 
     @Override
     public void removeFriend(int userId, int friendId) {
-        UserDB user = userRepository.getUserById(userId);
-        userRepository.getUserById(friendId);
+        if (userRepository.getUserById(userId).isEmpty()) {
+            throw new NotFoundException(String.format("Пользователь id = %d не найден", userId));
+        }
+        UserDB user = userRepository.getUserById(userId).get();
+        if (userRepository.getUserById(friendId).isEmpty()) {
+            throw new NotFoundException(String.format("Пользователь id = %d не найден", friendId));
+        }
+        user.setFriends(userRepository.getAllFriendOfUser(userId));
         userRepository.removeFriend(userId, friendId);
         user.getFriends().remove(friendId);
     }
 
     @Override
     public List<UserDto> findAllFriends(int userId) {
-        UserDB user = userRepository.getUserById(userId);
+        if (userRepository.getUserById(userId).isEmpty()) {
+            throw new NotFoundException(String.format("Пользователь id = %d не найден", userId));
+        }
+        UserDB user = userRepository.getUserById(userId).get();
+        user.setFriends(userRepository.getAllFriendOfUser(userId));
         return user.getFriends().stream()
-                .map(userRepository::getUserById)
+                .map(friendId -> {
+                    if (userRepository.getUserById(friendId).isEmpty()) {
+                        throw new NotFoundException(String.format("Пользователь id = %d не найден", friendId));
+                    }
+                    UserDB friend = userRepository.getUserById(friendId).get();
+                    friend.setFriends(userRepository.getAllFriendOfUser(friendId));
+                    return friend;})
                 .map(this::mapToUser)
                 .map(UserMapper::mapToUserDto)
                 .collect(Collectors.toList());
@@ -120,11 +152,25 @@ public class InDBUserService implements UserService {
 
     @Override
     public List<UserDto> findCommonFriends(int userId, int otherUserId) {
-        UserDB user = userRepository.getUserById(userId);
-        UserDB otherUser = userRepository.getUserById(otherUserId);
+        if (userRepository.getUserById(userId).isEmpty()) {
+            throw new NotFoundException(String.format("Пользователь id = %d не найден", userId));
+        }
+        UserDB user = userRepository.getUserById(userId).get();
+        user.setFriends(userRepository.getAllFriendOfUser(userId));
+        if (userRepository.getUserById(otherUserId).isEmpty()) {
+            throw new NotFoundException(String.format("Пользователь id = %d не найден", otherUserId));
+        }
+        UserDB otherUser = userRepository.getUserById(otherUserId).get();
+        otherUser.setFriends(userRepository.getAllFriendOfUser(otherUserId));
         return user.getFriends().stream()
                 .filter(friendId -> otherUser.getFriends().contains(friendId))
-                .map(userRepository::getUserById)
+                .map(friendId -> {
+                    if (userRepository.getUserById(friendId).isEmpty()) {
+                        throw new NotFoundException(String.format("Пользователь id = %d не найден", friendId));
+                    }
+                    UserDB friend = userRepository.getUserById(friendId).get();
+                    friend.setFriends(userRepository.getAllFriendOfUser(friendId));
+                    return friend;})
                 .map(this::mapToUser)
                 .map(UserMapper::mapToUserDto)
                 .collect(Collectors.toList());
